extract to base class filter menu
