+++ app/src/main/java/org/stepic/droid/ui/fragments/MyCoursesFragment.java
@@ -1,14 +1,7 @@
 package org.stepic.droid.ui.fragments;
 
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.widget.Toast;
-
 import com.squareup.otto.Subscribe;
 
-import org.stepic.droid.R;
 import org.stepic.droid.base.CoursesDatabaseFragmentBase;
 import org.stepic.droid.events.courses.FailCoursesDownloadEvent;
 import org.stepic.droid.events.courses.FailDropCourseEvent;
@@ -27,27 +20,6 @@ public class MyCoursesFragment extends CoursesDatabaseFragmentBase {
 
     public static MyCoursesFragment newInstance() {
         return new MyCoursesFragment();
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setHasOptionsMenu(true);
-    }
-
-    @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        inflater.inflate(R.menu.my_courses_menu, menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_filter_menu:
-                Toast.makeText(getContext(), "Hello, it is fitler", Toast.LENGTH_SHORT).show(); // FIXME: 19.08.16 FIX THIS
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
     }
 
     @Override
